Fix paths so all URLs resolve on GitHub Pages
Another attempt at last commit. For this had to move all HTML files to the root but can keep the JS in the subdirectories.

Since my site is on GitHub Pages with a custom domain, this repos' site is actually in a subdirectory so these links try to resolve from the root but fail. The `./` prefix should hopefully resolve that.