Use absolute value for screen edge collision
No need to flip the value, since we have the right conditionals for the screen edges, we just need the absolute value. This is best because if we flip the value but the object can't get back within the edges, then the value flips back. This means the object gets stuck on the edge and the coordinate would just flip from negative to positive and back again, over and over. The absolute value in all 4 edges solves this so the object always comes back into view.